+++ src/main/java/br/com/possoler/api/api_posso_ler/api/respondeai_api/controller/ExerciseController.java
@@ -1,8 +1,7 @@
 package br.com.possoler.api.api_posso_ler.api.respondeai_api.controller;
 
 import br.com.possoler.api.api_posso_ler.api.respondeai_api.dto.request.ExerciseRequestDTO;
-import br.com.possoler.api.api_posso_ler.api.respondeai_api.interfaces.RespondeAiClient;
-import org.springframework.beans.factory.annotation.Qualifier;
+import br.com.possoler.api.api_posso_ler.api.respondeai_api.service.ExerciseService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,10 +11,10 @@ import javax.validation.Valid;
 @CrossOrigin(origins = "*")
 public class ExerciseController {
 
-    private final RespondeAiClient respondeAiClient;
+    private final ExerciseService exerciseService;
 
-    public ExerciseController(@Qualifier("getExercise") RespondeAiClient respondeAiClient) {
-        this.respondeAiClient = respondeAiClient;
+    public ExerciseController(ExerciseService exerciseService) {
+        this.exerciseService = exerciseService;
     }
 
     @PostMapping("${respondeai-api.endpoint.getExerciseData}")
@@ -23,7 +22,7 @@ public class ExerciseController {
         @RequestHeader(name = "Authorization") String token,
         @RequestBody @Valid ExerciseRequestDTO payload
     ){
-        Object response = respondeAiClient.getData(payload.getItemId(), token);
+        Object response = exerciseService.getExerciseData(payload.getItemId(), token);
         return ResponseEntity.ok().body(response);
     }
 }
